Enum compose_yml::v2::CommandLine

pub enum CommandLine {
    ShellCode(RawOr<String>),
    Parsed(Vec<RawOr<String>>),
}

A command line to be executed by Docker.

Variants
ShellCode(RawOr<String>)

A command-line specified as unparsed shell code.

Parsed(Vec<RawOr<String>>)

A pre-parsed command-line. This may actually be empty for fields like command, so we don't try to enforce a minimal length, even if other fields like entrypoint supposedly want at least one entry.