Struct compose_yml::v2::GitUrl

pub struct GitUrl { /* fields omitted */ }

URL of a Git repository. Git repositories may be specified as either ordinary http or https URLs, or as scp-style remote directory specifiers.

One of the goals behind this class is to be able to use it as an "enhanced string", much like PathBuf, that can be passed to various APIs using conversion via AsRef and From. So we implement plenty of conversions, plus Ord so we can be used as a key in a BTreeMap.

Methods

impl GitUrl

fn should_treat_as_url<S: AsRef<str>>(s: S) -> bool

Would docker-compose interpret this string as a URL? We check against a list of known prefixes that trigger URL intepretation in docker-compose.

fn new<S: Into<String>>(url: S) -> Result<GitUrl>

Create a GitUrl from the specified string. Will only return an error if should_treat_as_url returns false.

fn to_url(&self) -> Result<Url>

Convert a GitUrl to a regular url::Url object.