Struct compose_yml::v2::Network

pub struct Network {
    pub driver: Option<RawOr<String>>,
    pub driver_opts: BTreeMap<String, RawOr<String>>,
    pub external: Option<ExternalNetwork>,
    // some fields omitted
}

A service which will be managed by docker-compose.

Fields
driver: Option<RawOr<String>>

The name of the network driver to use.

driver_opts: BTreeMap<String, RawOr<String>>

Options to pass to the network driver.

external: Option<ExternalNetwork>

Mutually-exclusive with all other options.

TODO LOW: We could represent Network and ExternalNetwork as some kind of enum, but that might break in the future if things get more complicated. For now, we're sticking close to the file format even if it makes things a bit less idiomatic in Rust.