Struct compose_yml::v2::VolumeMount

pub struct VolumeMount {
    pub host: Option<HostVolume>,
    pub container: PathBuf,
    pub permissions: VolumePermissions,
    // some fields omitted
}

A volume associated with a service.

Fields
host: Option<HostVolume>

If this volume is external to the container, where should we find it?

container: PathBuf

Where should we mount this volume in the container? This must be an absolute path.

permissions: VolumePermissions

What should the permissions of this volume be in the container?

Methods

impl VolumeMount

fn host<P1, P2>(host: P1, container: P2) -> VolumeMount where P1: Into<PathBuf>, P2: Into<PathBuf>

Map a host path to a container path.

use compose_yml::v2 as dc;
dc::VolumeMount::host("./src", "/app");Run

fn named<S, P>(name: S, container: P) -> VolumeMount where S: Into<String>, P: Into<PathBuf>

Map a named volume to a container path.

use compose_yml::v2 as dc;
dc::VolumeMount::named("pgvolume", "/app");Run

fn anonymous<P>(container: P) -> VolumeMount where P: Into<PathBuf>

An anonymous persistent volume which will remain associated with this service when it is recreated.