Struct compose_yml::v2::File

pub struct File {
    pub services: BTreeMap<String, Service>,
    pub networks: BTreeMap<String, Network>,
    // some fields omitted
}

A docker-compose.yml file.

Fields
services: BTreeMap<String, Service>

The individual services which make up this app.

networks: BTreeMap<String, Network>

The networks used by this app.

Methods

impl File

fn read<R>(r: R) -> Result<Self> where R: Read

Read a file from an input stream containing YAML.

fn write<W>(&self, w: &mut W) -> Result<()> where W: Write

Write a file to an output stream as YAML.

fn read_from_path<P>(path: P) -> Result<Self> where P: AsRef<Path>

Read a file from the specified path.

fn write_to_path<P>(&self, path: P) -> Result<()> where P: AsRef<Path>

Write a file to the specified path.

fn inline_all(&mut self, base: &Path) -> Result<()>

Inline all our external resources, such as env_files, looking up paths relative to base.

fn make_standalone(&mut self, base: &Path) -> Result<()>

Convert this file to a standalone file, with no dependencies on the current environment or any external files. This does not lock down the image versions used in this file.